refactor: component 폴더의 pages폴더 page와 동일 구조로 변경
