+++ src/components/pages/diary/diary/Content.tsx
@@ -1,9 +1,9 @@
-import EmotionTag from "../../common/EmotionTag";
+import EmotionTag from "../../../common/EmotionTag";
 import HeartIcon from "../../../assets/svg/heart.svg?react";
 import { useEffect, useRef, useState } from "react";
-import { addLike, removeLike } from "../../../api/api";
+import { addLike, removeLike } from "../../../../api/api";
 import { useParams } from "react-router-dom";
-import Divider from "../../common/Divider";
+import Divider from "../../../common/Divider";
 
 interface ContentProps {
   date: string;
