+++ src/components/pages/diary/diary/ImageCarousel.stories.tsx
