+++ src/components/pages/diary/diary/ImageCarousel.tsx
@@ -1,5 +1,5 @@
 import { useState, useRef, useEffect } from "react";
-import { DiaryImage } from "../../../types/types";
+import { DiaryImage } from "../../../../types/types";
 import ImageAddFilter from "./ImageAddFilter";
 import defaultImage from "../../../assets/images/defaultImage.png";
