+++ src/components/pages/diary/diary/NoDiary.tsx
@@ -1,8 +1,8 @@
 import { useNavigate } from "react-router-dom";
-import Appbar from "../../common/Appbar";
+import Appbar from "../../../common/Appbar";
 import non from "../../assets/icon/non.png";
-import Button from "../../common/Button";
-import ROUTE_PATH from "../../../constants/ROUTE_PATH";
+import Button from "../../../common/Button";
+import ROUTE_PATH from "../../../../constants/ROUTE_PATH";
 
 interface NoDiaryProps {
   date: string;
