+++ src/components/pages/main/album/SearchBar.stories.tsx
