+++ src/components/pages/user/stat/EmotionBarChart.tsx
