+++ src/components/pages/write/ReviewContent.tsx
@@ -1,5 +1,5 @@
 import EmotionTag from "../../common/EmotionTag";
-import ImageCarousel from "../diary/ImageCarousel";
+import ImageCarousel from "../diary/diary/ImageCarousel";
 import HeartIcon from "../../../assets/svg/heart.svg?react";
 import { DiaryImage } from "../../../types/types";
